I have create a class called Color Generator.
This class will reserve all colors.

so far using Test Driven Development TDD I have made an array of colors.

This array is testable for its length to be 32,768 units.

This array is testable for the increment value.

It helps me to make sure such a big array is constructed perfectly.